moved nuke utils back into utils
