+++ src/__init__.py
@@ -4,9 +4,8 @@ from __future__ import print_function
 import sys
 import logging
 
-
-from . import logger
-from . import resources
+from src import logger
+from src import resources
 
 LOGGER = logging.getLogger('ProfileInspector.init')
 try:
