+++ src/_nuke/__init__.py
@@ -1,2 +1,2 @@
-from ._nuke import *
-from .nuke_util import *
+from .fake_nuke import *
+# from .nuke_util import *
