+++ src/util/__init__.py
@@ -1,2 +1,3 @@
 from .time_formatter import TimeFormatter
 from .widget_color import widget_color
+from .nuke_util import *
